requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
gunicorn==21.2.0
requests
gspread
oauth2client